Tests/LibWeb/Text/expected/abspos-relayout-fixed-cb-toggle.txt

before-toggle partial=1 full=0
after-toggle partial=0 full=1
fixed-before-toggle=90,120,50,20
fixed-after-toggle=30,40,50,20
PASS
